fix attachments filtering; fix select category; add placeholder and validations

+++ sources/dronesapp-web/angular/src/app/components/forms/f-project-add/f-project-add.component.ts
@@ -36,6 +36,7 @@ export class FProjectAddComponent implements OnInit {
   ];
   isNotAcceptedFormat: boolean = false;
   isLimitReached: boolean = false;
+  isSubmitted: boolean = false;
 
   date = {
     start: moment(),
@@ -43,22 +44,16 @@ export class FProjectAddComponent implements OnInit {
   };
 
   services: NormalizedServiceModel[] = [];
-  service: string = null;
 
   categories: CategoryModel[] = [];
-  category: string = null;
 
   countries: CountryModel[] = [];
-  country: string = null;
 
   states: StateModel[] = [];
-  state: string = null;
 
   budgets: BudgetModel[] = [];
-  budget: string;
 
   durations: DurationModel[] = [];
-  duration: string;
 
   paidOptions: PaidOptionModel[] = [];
   paidOptionsChecked: boolean[] = [];
@@ -86,7 +81,7 @@ export class FProjectAddComponent implements OnInit {
 
     this.uploader.onSuccessItem = (item, response, status, headers) => {
       console.log('onSuccessItem', response, item);
-      const attachment = this.createAttachment(JSON.parse(response)['url'], item.file.name);
+      const attachment = this.createAttachment(JSON.parse(response)['url'], item.file.name); // TODO: handle same names of files
       this.formData.attachments.push(attachment);
       this.uploader.clearQueue();
       return {item, response, status, headers};
@@ -136,22 +131,22 @@ export class FProjectAddComponent implements OnInit {
         max: null,
         min: null,
         order: null,
-        title: ''
+        title: null
       },
       duration: {
         id: null,
         max: null,
         min: null,
         order: null,
-        title: ''
+        title: null
       },
       service: {
         id: null,
-        name: '',
+        name: null,
         category: {
           order: null,
           id: null,
-          name: '',
+          name: null,
         }
       },
       title: '',
@@ -160,7 +155,12 @@ export class FProjectAddComponent implements OnInit {
       location: {
         country: {
           id: null,
-          name: ''
+          name: null
+        },
+        state: {
+          code: null,
+          id: null,
+          name: null
         },
         address: '',
         city: '',
@@ -180,7 +180,6 @@ export class FProjectAddComponent implements OnInit {
       this.initDate();
     }
   }
-
 
 
   private getData() {
@@ -247,7 +246,6 @@ export class FProjectAddComponent implements OnInit {
 
     this.setServiceToPostData(service);
     this.categories = service['category'];
-    this.category = null;
   }
 
   private setServiceToPostData({id, name, order}) {
@@ -262,7 +260,6 @@ export class FProjectAddComponent implements OnInit {
 
       this.categories = service['category'];
     }
-
   }
 
   selectBudget(title: string) {
@@ -290,7 +287,6 @@ export class FProjectAddComponent implements OnInit {
       this.getListOfStates();
       this.clearState();
     } else {
-      this.state = '';
       delete this.formData.location.state;
     }
   }
@@ -327,15 +323,25 @@ export class FProjectAddComponent implements OnInit {
   }
 
   checkCountry() {
-    return this.country === 'United States';
+    return this.formData.location.country.name === 'United States';
   }
 
   postProject(e, form) {
     e.preventDefault();
+    this.isSubmitted = true;
 
+    if (!form.valid) {
+      return;
+    }
+
+    // console.log('project data to save:', this.formData);
     if (this.isEditForm) {
       return this.projectService.updateProject(this.formData)
-        .subscribe();
+        .subscribe(
+          // res => {
+          //   console.log('updated project:', res);
+          // }
+        );
     } else {
       this.formData.budget.confirmationValid = true;
       this.formData.confirmationValid = true;
@@ -344,7 +350,11 @@ export class FProjectAddComponent implements OnInit {
       });
 
       return this.projectService.postProjects(this.formData)
-        .subscribe()
+        .subscribe(
+          // res => {
+          //   console.log('saved new project:', res);
+          // }
+        )
     }
   }
 
@@ -409,15 +419,18 @@ export class FProjectAddComponent implements OnInit {
   }
 
   deleteAttachment(id) {
-    this.projectService.deleteAttachment(id)
-      .subscribe(
-        () => {
-          this.formData.attachments = this.formData.attachments.filter(attach => attach.id !== id);
-        },
-        err => {
-          console.log('delete attached file error', err);
-        }
-      );
-
+    if (this.isEditForm) {
+      this.projectService.deleteAttachment(id)
+        .subscribe(
+          () => {
+            this.formData.attachments = this.formData.attachments.filter(attach => attach.id !== id);
+          },
+          err => {
+            console.log('delete attached file error', err);
+          }
+        );
+    } else {
+      this.formData.attachments = this.formData.attachments.filter(attach => attach.id !== id);
+    }
   }
 }
